Linux & Xbox 360 Controller
Updated for the Microsoft Xbox 360 controller input under Linux (openSUSE). Since the Linux driver for the Xbox 360 controller is a port of the Mac driver, it is possible that this modification works as well on Mac (further investigation needed on this).
Added suggested code for the "BigButton" in comment which is possibly the Guide button. (Key: "Suggestion: Xbox Guide button <=> BigButton")
Added checking gamepad's name to identify Microsoft Xbox controller.