Adds strict Dispose behavior to Game
Game now mimics what the XNA Game object does after it is Disposed.
The effects of this are not evident with the XNA framework itself, since
Game.Run is blocking.  But for iOS and MacOS, the strict disposal checks
make it much easier to debug accidental disposal of the Game object
while it is still in use.