Fixes a runtime JIT that was occuring on iOS
The AOT compiler was not able to resolve the List<AddJournalEntry<T>>
constructor, so AddJournalEntry was moved outside
SortingFilteringCollection, and the generic parameter was removed.  This
is not ideal, but performance shouldn't suffer too much, since we happen
to know that any boxing that would have occurred to IUpdateable or
IDrawable has already occurred before they are stored as Object.