Fixes iOS app startup in non-Portrait orientations
iOSGameViewController now does its best to divine the best frame
rectangle for the view it manages.  iOSGamePlatform forces the current
GraphicsDevice's Viewport to Window.ClientBounds in BeforeInitialize.
This is not ideal.  GraphicsDevice itself should correctly query the
view for its bounds, but the changes needed to support that are
extensive and no doubt conflict significantly with develop3d.