iOS - Texture2D - Change FromFile so you can load from Documents folder
The FromFile method currently calls UIImage.FromBundle which takes advantage of automatically loading @2x images for high resolution retina devices.  However it can only load files from the Content folder.  The UIImage.FromFile method allows you to load graphics from the document folder as well as the Content folder, however it does not automatically cope with loading @2x images like the FromBundle method does.  Therefore this change will check the path we are trying to load - if it is from the Content folder then we continue to call UIImage.FromBundle, however if it is from anywhere else then we will call UIImage.FromFile.