Texture2D - Support for SaveAsPng and SaveAsJpeg
Added the methods SaveAsPng and SaveAsJpeg.

Both of these methods (and the CreateRGBImageFromBufferData method that they call) are taken from the MacOS MonoGame implementation with a very minor change to the CGImageAlphaInfo declaration to ensure the correct Alpha channel is stored and to ensure the Red and Blue are in the correct order.

This commit also ensures that the method signature matches the XNA API