Remove unused XnaGameWindowImpl.cs
This file existed both in Android and PSSuite folders. It was included in the MG Android csproj but was not used by any other code. Can only assume it was replaced by AndroidGameWindow.cs. The "Impl" suggested it might be related to unit tests, but I checked Kungfubanana's MonoGame-Tests repo and it also didn't use this file.