Relax some GL error checking to get Mac running again
Note that this is not the ideal and full fix. The setting of the
viewport is due to Game.Initialize() calling
GraphicsDevice.ApplyChanges() and this call has a comment that says
something to the effect of "TODO: we shouldn't need to do this". Rather
than go down the rabbit warren of changing that right now, I thought it
prudent to simply relax the error checking to get it running and worry
about correctness and initialisation order later.