Fix for a bug when trying to load a Song that's right inside Content's folder.
If you put a Song (wmb + xnb) inside the Content's folder, the CurrentAssetDirectory is used with the ContentManager.RootDirectory at SongReader's Read method, merging them and causing it to use the wrong path. I've tested this only on Windows8, so probably need to test this on other platforms too to make sure this is not a Windows8 bug only.