Fix GL related exception on Android after multiple resumes
The issue arose from the static list `GraphicsResource.resources` being cleared when the devices was reset. During the second reset, the shader resources weren't in the list so they couldn't be invalidated. Since most of these resources would survive the reset, they shouldn't be cleared. Instead the list is pruned of expired (ie garbage collected) weak references.

Also improved a comment and tightened the GL error checking from logging to exceptions in ShaderProgramCache. It seems that if ever these errors were logged, the program could not recover and it would fail subsequent error check anyway.