Android: Fix for black screen when turning screen on just after an auto screen off
If the user turns the screen on just after it has automatically turned off, the keyguard will not have had time to activate and the ActionUserPreset intent will not be broadcast. We need to check if the lock is currently active and if not re-enable the game related functions. http://stackoverflow.com/questions/4260794/how-to-tell-if-device-is-sleeping