Avoid adding mipmaps for non power of two textures on iOS
So far only tested on an iPad 1. This is a potential workaround for
issue #1028. Strictly speaking, one should not provide mipmaps on
non-pow2 textures if the target platforms don't support it. However,
since MonoGame is highly cross platform, and other platforms do support
it (eg Android), the best solution may be to simply ignore the mip maps
and log a debug message.