Implement GetData() for IndexBuffer & VertexBuffer on WinRT
It's quite common to use the standard importers to get a Model and then
use a GetData()  in order to read the Vertices/Indices and manipulate
them.
I use some  code from SharpDX.Toolkit as quide.
In short one has to copy the static VertexBuffer into a temporary buffer
which we can then read.