Make OpenGL vertex shader dirty on viewport change
The OpenGL version of MonoGame uses an uniform named "posFixup"
to offset any rendering by a fraction of the viewport size. This
is required because DirectX coordinates refer to pixel centers while
OpenGL coordinates refer to pixel corners.

Whenever the render target is changed, the vertex shader is marked
dirty so that the posFixup uniform gets updated (the viewport
changes whenever the render target is changed).

This change ensures the posFixup uniform is updated whenever the
viewport changes, not only when it is changed through a render
target change.