Reset Texture's cached GL SamplerState when device is reset
The cached sampler state was being persisted between device resets. After a context lost event, the GL sampler states are reset to default but the cached state was not. This meant that if your game used a single non-default sampler state, after a device reset it would never get applied.