Add support for Keys.RightControl/Keys.RightShift
Right now keyboard depends on CoreWindow.KeyUp/KeyDown events.
The problem is that those events does not distinguish between
LeftControl/RightControl (same for Shift,etc) but report a generic
VirtualKey.Control.

In that case i check against CoreWindow.GetKeyState(...) and the _keys
array to update the state of such keys.