Removed unnecessary garbage generation from Model.Draw
Problem: Model.Draw method was generating considerable amount of unnecessary garbage by allocating new bone transform matrices arrays with every call to the method.

Solution: New static field was introduced which replaces the role of bone transforms array. This array is static in order to save memory. We do not need separate instance for each Model because the transforms are set always directly before drawing the Model. At each Draw method call, the array is checked if it is null and is long enough hold transforms for Model with highest number of bones.