Fix for iOS Orientation on iOS6+
When using Guide.BeginShowKeyboardInput and having both Landscape Left
and Right orientations enabled, this method was forcing the orientation
of the presented KeyboardInputViewController to be LandscapeRight, even
if the current orientation was different. Now the code tries to keep the
current orientation. If this is not possible, only then it will try to
force Right, then Left and then Portrait.