OUYA start button workaround
Due to the recent firmware update, the OnKeyUp and OnKeyDown events for KeyCode.Menu are both being sent in a single frame. This stops MonoGame's GamePad implementation from ever registering the Start button as being pressed.

This workaround can be removed if the OUYA firmware is ever updated to send these events in different frames.