DynamicSoundEffectInstance
Implementation of DynamicSoundEffectInstance based on XAudio2 of
SharpDX. It's currently only implemented in the WP project but
could/should be migrated to the Windows 8 project as well.
As I couldn't find any unit tests on the Audio part of MonoGame I
performed a manual test by using the library in a separate WP8 MonoGame
project. It ran as expected.