VertexElementColor R,G,B,A properties must be public.
I am not really sure what is the purpose of VertexElementColor in
VertexPositionColor.
XNA documentation states that it should have been a 'Color'. Also one
should be able to define it's own IVertex structure and they would
normally use a 'Color' to do so.

What I see is that Color defines Getter/Setter for storing those values
into a packed int. Is this really necessary? A struct is a struct after
all. Same effect could have been achieved with StructLayout/FieldOffset
attributes.

Anyways. For now R/B/G/A properties must be publicly accessible as they
are on 'Color'.