Implemented "Garbage-Free" GraphicsDevice.GetRenderTargets
Implemented property in GraphicsDevice to get the current Render Count
and overloaded GetRenderTargets to copy to an array you provide instead
of generating a new one (therefore, creating garbage).  Only problem is
that it is the user's responsibility to make sure the array they create
can fit the render contents or an exception will be thrown.