TextureCubeReader ignore existingInstance
TextureCubeReader ignore existingInstance. As a result of this on WP8
when the app is deactivated/activated the ReloadAsset fails to reload
the Cube.

I applied the same fix to Texture3DReader although I didn't test it. The
fix is quite simple and should work.

I tried to make the code resemble as much the style of the
Texture3DReader.Read(...)