Set GraphicsDevice.PresentationParameters.DepthStencilFormat better on Android.
This change makes it be set based on the GraphicsMode of the AndroidGameWindow,
which also sets the actual depth and stencil settings used.

It still may not be entirely accurate if we fall back on default parameters for
the framebuffer; this might cause it to drop to DepthMode.None even if
defaults allow a depth buffer.