Implement ARB and EXT flavors of FBOs
MonoGame is currently relying on GL_ARB_framebuffer_object without
checking for support. This means it will fail to run on systems that do
not support OpenGL 3.x+.

This patch implements a fallback to GL_EXT_framebuffer_object when
GL_ARB_framebuffer_object is not supported.

Note that we cannot use GL_EXT_framebuffer_object alone, since many
systems support the ARB flavor but not the EXT one.