Process the hold gesture even if no touch events are occurring.
Previously, new gestures were only checked for when _gestureEvents was non-empty. Android at least does not add to _gestureEvents while a touch location stays still, and so would not recognise a hold gesture unless and until the user moved the touch location (at least slightly) at some point after the appropriate time had passed. If the hold was released without moving, the hold gesture would not be processed at all. This caused the hold gesture to be unreliable unless the user deliberately performed a small motion after a pause to trigger it.

This commit causes gestures to be processed unconditionally of whether any new gesture events were received.