Work around OpenGL contexts not being immediately usable on OS X 10.7+
Recent OS X gives an InvalidFramebufferOperationExt error if you draw to the default framebuffer while its associated NSView is still not visible. This means we need to not draw until the window is visible. If the render target usage is set to DiscardContents (the default), we try to clear the default framebuffer during initialisation, when we apply the default render target, causing this error.

This is a lazy fix that should work reliably for us. A more complex but backwards compatible and consistent solution would be good if other people have this issue.