MGFX hashCode
Write a hash code of the shader/effect in the  Header.

File version up to 6 from 5 because the header has chaneged. I had to
make room for 4 more bytes.

Everything after the header (ConstantBuffers/Shaders/etc) has to be
written into a Memory Stream.
We then calculate a hash from this memory stream and write it to the
header. at the end the memory stream is written into the final stream.
This makes it possible to accept a non seekable stream (network?) but
use more memory.
The alternative would be to store the position, write a null value,
write the rest of the file and seek back to update the hash value.