Fix for ResolveRelativePath stripping drive char
As per issues #3064 and #3041, parsing a file path through the
ResolveRelativePath function was stripping out the drive letter. This
crash was especially prominent for the sound and video loaders.

This fix will strip out the platform-specific path separation character
if it exists and will not try to remove it if it doesn't.