FixMemoryLeak on SongReader
Return existingInstance instead of creating a new one.
I already have another PullRequest to exclude types such as Song and
SoundEffect from getting reloaded, so this fix will only be relevant in
case of custom XNBs with embedded Songs.
As always i make sure that all data is read from input so the stream
will move to the right position on return whether this is an
existingInstance or not.