Upgrade Android to OpenTK-1
Also added Property to Activity to control if the rendering is done on
the UI thread or not. By not rendering on the UI thread we save
a SyncContext.Send call in the underlying AndroidGameView.
This can introduce up to a 100ms delay in a Render call which
can cause a stutter.

By default we will leave RenderOnUIThread = true to maintain
backwards compatability.

The upgrade to OpenTK-1 was required as the fix for this
is not present in the OpenTK 0.9.9.3 version as it is there
for backwards compatabiltiy only.