Implement vertexStride in VertexBuffer.SetData for DirectX platform.
This code roughly mirrors the equivalent path in GetData, although it is
only used when vertexStride != elementSize, so it doesn't affect
existing usage. An alternative implementation (which I actually tried
first), which avoids the need for a staging buffer, would call
UpdateSubresource once for each element, but I guess that would be very
slow for large arrays.

Note that this only implements the vertexStride parameter for
non-dynamic VertexBuffers (which is actually harder than it would be for
dynamic VertexBuffers, since you wouldn't need the staging buffer for a
dynamic VertexBuffer).

I changed Game.RunOneFrame() to Game.Run() for these tests, because
Game.RunOneFrame() doesn't seem to call Draw() when running tests in
XNA, so the tests gave false positives.