Fixed parameterized Next overloads
In the process, I replaced the Remap functions with in place modulo
operations. This does NOT introduce bias:
The scaling-based Remap function distributed more commonly occuring
values over the range based on rounding error. The modulo operation has
the same number of more common numbers, just in sequence, beginning at
the low end of the range.