[SkinnedModel] struct Keyframe
Convert Keyframe to struct.
Instead of allocating a large number of Keyframes and add each one on a
List<Keyframe> we now allocate a single array of structs.
The idea is to keep a low heap complexity (number of live objects &
number of object references) for the GC.

Also this will improve loading times and reqired memory by a small
factor.