[pipeline] Added support for attaching the mono debugger to MGCB via the pipeline tool
One feature that was missing in the pipeline tool was the ability to
debug your extensions via the UI. This commit adds that support
by providing the appropriate command line arguments to mono
which will cause the app to pause and wait for the debugger to
attach. A suitable warning message is included to make sure
the developer knows they are in debug mode.