Added support for MSBuild.Content.Builder on MacOS
In order to support the Content.Builder.targets on
MacOS we firstly needed to install the Tooling and
Assemblies in the right location. In this case
/Library/Frameworks/MonoGame.framework/vX.X.X.X

Symbolic links are required to get everything working
right. In this case we need a v3.0 link in
/Library/Frameworks/MonoGame.framework but we also need
some links in the Mono.framework as well.

We also need to add an Imports element to the csproj files.
Now the current MonoDevleop/Xamarin Studio templating system
does not support doing this directly. So a new MSBuild
processing class has been added to the addin. This class
will check if the project is a "MonoGame" project (is references
MonoGame.Framework.dll or has a MonoGamePlatform element) and
inject the required data into the csproj.

The templates have been left referencing the Nuget packages
for now.

Modified the ContentBuilder targets file to Batch its items
this is to get round issues when running under xbuild on Mac.

Updated templates to use a new MonoGameInstallDirectory property
rather than $(MSBuildProgramFiles32) so that we can use the
same paths for references.

It also adds the followiing to the projects

```
<MonoGameInstallDirectory Condition="'$(OS)' != 'Unix' ">$(MSBuildProgramFiles32)</MonoGameInstallDirectory>
<MonoGameInstallDirectory Condition="'$(OS)' == 'Unix' ">$(MSBuildExtensionsPath)</MonoGameInstallDirectory>
```

So that we can resolve the directory correctly on both Windows
Mac and Linux