[pipeline] El Capitan did some weird stuff with the $PATH and /usr/bin
Because of these changes mono can no longer be created in /usr/bin
as its read only. So the mono project have moved that exe to
/usr/local/bin.. which makes sense..

But /usr/local/bin is ONLY included in the path for the terminal/shell
and NOT for Applicaitons. So we end up in a situation where calls
to mono don't work. So the fix is to search for the locations in
the Pipeline tool AND the Build .targets.