Fix an issue where the user is unable to change the app Icon from default MonoGame one.
Because we explicitly use the Embedded Icon, unless a user builds
the library and changes it they cannot change the icon of the game.
While this might be nice from a MG point of view for marketing ;)
its not really desireable.

This commit adds support for looking for an "Icon.ico" file in
the applicaiton embedded resources first, then default to the
MonoGame one if needed.