Rename Test Class and update font finding procedure
Rename Pipeline test class "EnumComformingTest.cs" to
"PipelineEnumConformingTest.cs", avoids problem with the tests report
Link the new procedures to find a font in the following order:
1- Using SharpFont to read the each font metadata.
2- Platform specific, for Linux is untouched, for Windows it looks in
the registry
3- Filename matching
Modify exception handling to avoid code references inside in the
pipeline log in case the procedure fails.
Add extra information to the log to identify possible problems with the
procedure only in debug compilations.
Various fixes for edge cases