First Pass
Modify ImportGlyph signature to receveive two additional parameters to
tell if the font has to be emboldened and/or italiced.
Add method Embolden to add a faux bold effect.
Add method Italice to add apply a faux italic effect
Note that because this does not touch FontDescriptionStyle, it will
always embold the font