UWP Target Includes Rooted File Access
In addition to TitleContainer file system access, UWP apps fully support
read and write access to ApplicationData.Current.LocalFolder and
KnownFolder directories (Music Library, Pictures).  Not sure it is the
role of MonoGame to restrict access to the TitleContainer subfolder
exclusively if a rooted path is requested on the the WINRT\UWP platform.
If access to a rooted folder is invalid an access denied exception is
thrown for UWP platform targets (Windows 8, Windows 10, Windows Phone,
XB1).  This proposed code change adds WINRT to the list of platform
targets which allow rooted file system access.

Updated to use correct preprocessor macro, WINDOWS_UAP instead of WINRT.