Fix SDL Game Window Resize
For some reason SDL seems to sometimes report multipe resize
events with the same width and height. This can cause problems
for game if they are doing things like creating render targets
in the event.

So we need to make sure we only call the event when we actually change
the width or height.