Fix Mouse updating when moving the Window
When you click and hold the title bar of the window
and move it around we are still updating the X/Y
location of the Mouse. This does not happen in XNA
and it makes things look a bit weird as far as the
user is concerned.

This commit just moves the updating of the X/Y position
into the check that the mouse has Focus. This fixes the
issue since when we click on the Title bar the Mouse does
NOT have focus.