Update SDL Platform to Enable/Disable Screensaver of it loose/gain focus.
We really shouldn't be Disabling the screen saver completely.
If the user switches to another app the screen saver should be
re-enabled. As a result we also need to make sure we re-disable
it if we get focus again.

Additionally we need to make sure the Cursor is set because sometimes
on MacOS it looses the setting when switching windows