Added Localized Resource Processing and Loading.
When localizing games you sometimes need to load localized
assets as well as text. This commit includes the sample
XNA processors and ContentManager code into the core of the
MonoGame system. This means that people no longer need to
implement this stuff themselves.

Usage:

Define all of the "string" resources in standard .resx resource
files. Add these to the .spritefont file

```
   <ResourceFiles>
      <Resx>..\Strings.resx</Resx>
      <Resx>..\Strings.da.resx</Resx>
      <Resx>..\Strings.fr.resx</Resx>
      <Resx>..\Strings.ja.resx</Resx>
      <Resx>..\Strings.ko.resx</Resx>
    </ResourceFiles>
```

Change the .spritefont over to use the LocalizedFontProcessor.
Load the font as normal.

For other localized assets you can make use of the new

    ```Content.LoadLocalized<T>```

method. This will attempt to load assets which contain a culture
extension e.g

	Texture1.fr.xnb
	Texture1.en-gb.xnb

Note either the Name or the TwoLetterISOLanguageName can be used.