Implement effect building for GLFX
This is the start of the effect building implementation for the new GLFX
platform. Implements replacing functions with whitespace and passing entry
point functions for techniques to the optimizer. Not sure how we should
reassemble the file with the optimized code at this point. There will be
duplicate functions and uniforms in the result so either we just stitch
the results from the optimizer together and have some metadata at the
start of the file for the range of characters that a shader consists of or
we have some kind of system that lets us include or exclude parts
of the shader when compiling. Not sure if glsl compilers also do another
preprocessing step, but that's worth investigating.