[WP8] fix NullReferenceException on first draw.
The initial values of preferredBackBufferWidth/Height are 0 and it tries
to create a DepthStencil backbuffer of 0x0.
Gameloop starts in an invalid state and the first call to Game.Draw()
will fail with NullReferenceException on something like
`GraphicsDevice.Clear(Color.CornflowerBlue);`.

For whatever reason a simple new project recovers from this error
without problem and continues with a new DepthStencil backbuffer on the
next Draw().