Fallbacks if the preferred render target formats are not supported
DirectX 11 (DXGI 1.1) does not support 16-bit surface formats Bgr565, Bgra5551 or Bgra4444 for render targets.  DirectX 11.1 (DXGI 1.2) does support these surface formats for render targets.
https://msdn.microsoft.com/en-us/library/windows/desktop/ff471325(v=vs.85).aspx

This change implements a check for the support of the requested format and falls back to `SurfaceFormat.Color` and `DepthStencilFormat.Depth24Stencil8` if the requested formats are not supported.