Working on the runtime
The SpriteEffect shader actually loads with these changes. The matrix is
also passed correctly. There are issues with the sampler and with the
color of the vertices, but this is a good start :)

There are a few 'hacks' introduced in this PR to get the GLFX format
working without changing the 2MGFX format. Look for #if GLFX
preprocessor macros.