Fix Shader tests
The Rainbow test is still failing, it is not drawing the sprite
at all. We use 'any' in the DirectX shader, I tried replacing
that by testing alpha, but that did not fix the issue for OpenGL,
it still did not render anything. Not sure what to the issue is/how
to fix.