DesktopGL was not using back buffer and depth buffer formats
OpenGL was being initialized with default OpenGL formats, which does not include a stencil buffer.  SDL2 requires the SDL GL attributes to be set before the window is created.  The window creation has been delayed to just prior to the graphics device being created and the setting of the SDL GL attributes was moved earlier to just prior to the window being created.  Unit tests were added for stencil clip.

Fixes #5441