[DesktopGL] Improve window initialization
This makes some improvements to how the GameWindow works:

- Multisampling is supported by the window on initilization. I create a
dummy window to create a temporary GraphicsContext to query for
multisampling support. I got some weird behaviour though, so would be
nice if someone can test that.
- The window position can be set before the native window is instantiated.
- The window title can be set before the native window is instantiated.
- Initializing in full screen mode no longer depends on the internal
'applyChanges' method in game. I #iffed out DESKTOPGL for that method so
we know it no longer needs it. One step closer to getting rid of it
completely.
- Stop listening for SDL Resized event. If it is raised, the
SizeChanged event is raised too, so listening to that is sufficient.
- The ClientSizeChanged event is only raised when a user resizes the
window (also while resizing, not just when resizing ends but there's no
way to distinguish the two with the SDL API), not when it is resized
through a back buffer size change.
- The window position is restored after exiting full
screen. This is not very stable because SDL raises a bunch of events
when entering/exiting full screen and we can't distinguish them from
when a user moves the window. Optionally we can check if window position
in the event is (0, 0) relative to the current display and assume its a full
screen transition and ignore it when it is.
- Window centering now also happens after resizing the backbuffer in
stead of just when exiting full screen. This is also not very stable
because of the previous issue. Sometimes we detect a window move when
it's because of a full screen switch and afterwards the window will no
longer be centered.

I also added some todo's/comments for more intricate stuff like
recreating the window when certain presentation parameters change.