Fix up some weird Effects crashes.
One of the samples I am porting hit this issue where the
EffectParameter.SetValue (int) for Single/Int parameters
as well as for Scalar and Vector parameters.

There was also a problem with access the Element[i] array in
EffectParameter.SetValue(Single[]). In this case it was crashing
event though the Data object was correct. This commit
removes the use of Element in favour of Data which other Setters
use.