#if out MAC code + directory properties
I'm a bit fresh to C#, but how's this?

I don't think it makes sense for Mac to have `FindFontFileFromFontName`... it just needs to add search directories to the list. However, this whole chunk of fontName calculation could be helpfully extracted out...

With this commit, OS-specific logic for the fonts & local fonts directories are defined as applicable by get-only properties.

With this approach, `Process()` isn't cluttered with this directory-determining logic, plus less `#ifs`.