Fix GetBackBufferData for DX
Previously when the Color format was specified (R8G8B8A8) it was changed
to BGRA before creating the swapchain. I think the reason for it is that
flipping the buffer is slightly faster when using BGRA because the native
format is BGRA, so no swizzling is needed when flipping the buffers.
In any case, this is not a breaking change and it makes more sense to use the
requested buffer format. People that want to optimize this far can
just request a BGRA backbuffer to get the fast flipping benefits.