Reusable function for raising events
Before raising an event, you must first check the handler to see if it is null in order to check that an object is subscribed to handle the event.  This is best done if the event handler is stored in a local variable just in case the last subscriber unsubscribes in a different thread between the time the event handler is checked and the handler is actually invoked.  This reusable function makes it easy to raise events by storing a local copy of the handler and checking for null before invoking it.  Many classes already had a helper function for this.  This change refactors those into a single function.

There are some instances where more complex logic was being done after the event handler was checked.  In those scenarios, the code was mostly left alone except for the introduction of a local handler variable.