Fix hard full screen on startup and disposal
Hardware full screen failed when set in the game constructor because there
was no swapchain yet to get the output from. If there is no swapchain yet,
the current method gets the primary output by enumerating adapters and
getting the first, then getting the first output from that adapter.

When exiting a game while in full screen I sometimes got stuck with a
black screen. That was fixed by making sure we exit full screen when
disposing a GraphicsDevice.