Allow DirectX debug layers to be enabled via environment variable (#5786)
When using a release build of MonoGame (as is the case whenever using a prebuilt binary), there's no way of enabling the DirectX debug layers, even on systems with those debug layers installed.  To support this scenario, this updates device creation so that if the MONOGAME_ENABLE_DIRECTX_DEBUGGING environment variable is set to "true", then MonoGame will attempt to enable the debug layers if they are available.