Fixed a crash when a newer mono is installed.
If a user has a different version of mono installed on a Mac
the Pipeline tool crashes when building. This is because the
mscorlib in the pipeline tool directory (which is used by default)
is not the same as the one the `mono` executable was built against.

This commit adds the setting of the `MONO_PATH` environment
variable to ensure that `mono` picks up the system version rather
than the one in the app itself.