Fix a Font issue on MacOS #5900 Bump Dependencies to bring in windows specific SharpFont assembly.
This is all down to mono now running in 64bit (we think). So the
old code we had no longer worked. The Dependencies include specific
versions of SharpFont for Mac and Linux and a seperate one for
Windows. The Mac/Linux ones support AnyCPU but the windows is
x64 specific. Since we require x64 for Windows content this
should not be a problem.