Fix a crash when building content in xbuild  (#5897)
* PR #4997 Introduced a crash in xbuild. This is because of a bug in
xbuild where `<ItemGroup>` with Metadata are not supported. Instead
you need to use `CreateItem`.

This commit fixes that issue.

There are also other issues which need to be resolved with #4997.

That said this commit also adds a new unit test for the content system.
There is a sample project which we build as part of the unit test.
We check the output to make sure it is in the right place. This test
should be expaned on to include tests for Shared Projects and other
senarios that users use. It should allow use to catch issues such as
what happened in #4997 at the point the PR is opened.

* Updated Definiton to copy MacOS files on Mac

* Ignore a couple of Image loading tests as the formats are not supported

* Enable Tests on MacOS

* Fix Tests running on MacOs

* Rmeove the use of mdtool. It is obsolete

* More fixes for the Content Builder Targets

* More fixes for the Content Builder Targets

* More fixes for the Content Builder Targets

* More fixes for the Content Builder Targets

* More fixes for the Content Builder Targets

* Added MonoGame.Build.Tasks to the project.

* Added some documentation

* Fixed some errors in the docs

* PR #4997 Introduced a crash in xbuild. This is because of a bug in
xbuild where `<ItemGroup>` with Metadata are not supported. Instead
you need to use `CreateItem`.

This commit fixes that issue.

There are also other issues which need to be resolved with #4997.

That said this commit also adds a new unit test for the content system.
There is a sample project which we build as part of the unit test.
We check the output to make sure it is in the right place. This test
should be expaned on to include tests for Shared Projects and other
senarios that users use. It should allow use to catch issues such as
what happened in #4997 at the point the PR is opened.

* Updated Definiton to copy MacOS files on Mac

* Ignore a couple of Image loading tests as the formats are not supported

* Enable Tests on MacOS

* Fix Tests running on MacOs

* Rmeove the use of mdtool. It is obsolete

* More fixes for the Content Builder Targets

* More fixes for the Content Builder Targets

* More fixes for the Content Builder Targets

* More fixes for the Content Builder Targets

* More fixes for the Content Builder Targets

* Fixed msbuild test on windows

* Fixed windows tests

* Fixed Task so we don't get an extra folder