Fix Duplicate/Weird Menus on MacOS
Eto Forms seems to default `IncludeSystemItems` to `All`.
As a result the deafult system menus like `File`, `Edit`, `Help`
etc are populated by default on MacOS.

Even though we set the property later they are not cleaned out.
So we end up with duplicates and messy menus.

So we subclass the `MenuBar` and set the property to our required
default first.