[OpenGL] Rework Android Framebuffer Support
OpenTK removal also removed some support for frame buffers
on older devices. The old code was a bit of a mess with
mutliple FrameBuffer classes each doing the same thing.

This commit reworks the code to have one OpenGL FramebufferHelper
and moves all the loading of OpenGL extensions into the OpenGL.cs
class where they belong.

As a result GraphicsCapabilites was reworked to use the `Extensions`
property on the `GL` class. We also needed to rework when the
OpenGL entry points are loaded so ensure we have a current context.
If we don't we cannot get the extension strings from opengl since
it required a current context.