Fix Pipeline tool to work regardless of Mono changes.
One of the problems we have with the Pipeline tool
at the moment on Mac is this. When running the MGCB.exe
using system `mono` we can end up in a situation where
the wrong versions of the system assemblies and mscorlib
are loaded. As a result some calls to the system will
fail because entry points are missing or are invalid.

This is an attempt to fix that by bundling the `mono`
we are built against inside the Pipeline.app package.
In theory this should help.

I tried setting `MONO_PATH` and various other mono related
flags but nothing worked.